refactor: refactor examples, fix config check
