+++ foyer-storage/src/store.rs
@@ -101,8 +101,8 @@ where
             catalog_shards: 64,
             admissions: vec![],
             reinsertions: vec![],
-            flushers: 4,
-            reclaimers: 4,
+            flushers: 1,
+            reclaimers: 1,
             clean_region_threshold: None,
             recover_concurrency: 8,
             compression: Compression::None,
